fix job board
